requirements.txt
fastapi
uvicorn
google-genai
google-adk
fastmcp
python-dotenv